models/icl/mobilenet/pyproject.toml
[tool.poetry]
name = "npu-examples[mobilenet]"
version = "0.1.0"
description = ""
authors = ["Zac <1221537+tezheng@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
npu-examples = { path = "../../..", develop = true }
numpy = "<2.0.0"
pillow = "^11.0.0"
datasets = "^3.2.0"
polars-lts-cpu = "^1.18.0"
olive-ai = "^0.7.1.1"
torchvision = "^0.20.1"

[tool.poetry.group.dev.dependencies]
torchinfo = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
